订单提交操作 将订单信息和订单一同提交 * mybatis返回主键的操作是 使用显式主键插入记录之后，将序列的当前值赋给对象的属性后调用函数显示出来 * mybaits调用oracle函数currval推测存在和会话一起的问题，可能会导致取到的currval不是正确的值，故不建议使用

+++ src/main/java/com/aspire/OnlineMeal/service/IUserOrderInfoService.java
@@ -3,14 +3,15 @@ package com.aspire.OnlineMeal.service;
 import java.math.BigDecimal;
 import java.util.List;
 
+import com.aspire.OnlineMeal.model.OrderInfo;
 import com.aspire.OnlineMeal.model.UserInfo;
 import com.aspire.OnlineMeal.model.UserOrderInfo;
 
 public interface IUserOrderInfoService {
 	//增加订单
-	public int addUserOrderInfoWithSelective(UserOrderInfo uoi) throws Exception;
+	public BigDecimal addUserOrderInfoWithSelective(UserOrderInfo uoi) throws Exception;
 	//获取主键
-	public int getPrimaryKey() throws Exception;
+	public BigDecimal getPrimaryKey() throws Exception;
 	//取消订单 将订单状态置为取消状态 不为删除订单操作
 	public int deleteUserOrderInfo(UserOrderInfo uoi) throws Exception;
 	//根据主键获取订单
